Restructure; move decompositions into new directory.
